Low-rank approximation and  matrix-free approach
This is the commitment with the implementations of a low-rank
approximation of the covariance matrix and the matrix-free approach.
The main implementations are kept in files starting with PCGA, however,
it was necessary to also make changes in other files.
The Microsoft visual studio project was updated to 2015. This version
includes a function which can make automatic indentation. Apparently
GitHub believe that changing the indentation is a change in a code line,
wherefore it looks like I have changed most of the code. I hope it is
possible to make GitHub ignore these types of changes so the real
changes can be inspected.

Here is a short description of the new implementations:
- After loading the input files and calculating the forward response of
the initial parameters the X-matrix is normalized using SVD.
- The covariance matrix is approximated by a low-rank approximation. For
this sampling from a Gaussian distribution is needed. Here fast Fourier
transformation techniques are used. This requires the unknown parameters
to be positioned in a regular grid.
- Using this low-rank approximation the matrix-free approach can be
used. This replace the need for computing the full Jacobian matrix.
- The Lambda-M system of equations is solved using the output of the
above points.
- When the convergence criterion is reached the posterior covariance is
computed. Either the full covariance or the diagonal elements are
computed.
- The structural optimization algorithm is available for the new
implementations.
- A Levenberg-Marquardt and line-search extension is available for the
new implementations.
- The demand of the unknown parameters to be placed in a regular grid
can be eased by using the function non-complete-grid. This function is
most functional if e.g. only few points are missing in the regular grid.

During the implementations the Intel Fortran compiler XE2016 was used
under a student license. I am not sure if I am allowed to publish the
compiled exe-file under this license, wherefore it is not uploaded here.

I saved everything in a Dropbox, wherefore I did not feel the need to
also secure my files at GitHub. It was not the intention to exclude
anybody from the ongoing implementations.

Minor changes are still needed, which will be made in the future. It is
however necessary to publish this version on GitHub due to my upcoming
examination in my master's project.